Concerns #110 - Moved PreProcessors from the core model
